Spawning, attacking, moving & rotating
Added better spawning of the enemies in the scene class.
Enemy attacks an given actor every given seconds when in range.
Enemy moves to given target and waits a given amount of time before going to the next one.
Rotate to face player setup is implemented.

NOTE: Moving is not working optimal, since the position cannot be retrieved from the position matrix of the enemy yet.